move utility functions over  word lengths #32
